* Headers/AppKit/NSPageLayout.h         * Source/NSPageLayout.m: Change int to NSInteger.         Patch by Sebastian Reitenbach <sebastia@l00-bugdead-prods.de>
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/gui/trunk@36092 72102866-910b-0410-8b05-ffd578937521

+++ Headers/AppKit/NSPageLayout.h
@@ -78,8 +78,8 @@ enum {
 //
 // Running the Panel 
 //
-- (int)runModal;
-- (int)runModalWithPrintInfo:(NSPrintInfo *)printInfo;
+- (NSInteger)runModal;
+- (NSInteger)runModalWithPrintInfo:(NSPrintInfo *)printInfo;
 #if OS_API_VERSION(GS_API_MACOSX, GS_API_LATEST)
 - (void)beginSheetWithPrintInfo:(NSPrintInfo *)printInfo
                  modalForWindow:(NSWindow *)docWindow
